pyproject.toml
[tool.poetry]
name = "mcp-tool-demo"
version = "0.1.0"
description = ""
authors = ["ramacharanreddy-feuji <ramacharan.r@feuji.com>"]
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
langchain-google-genai = "^2.1.9"
python-dotenv = "^1.1.1"
mcpconductor = {path = "../reqon-mcp-conductor", develop = true}
websockets = "^15.0.1"
cryptography = "^45.0.6"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-asyncio = "^1.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"